chore(oss) move to community
* rename package
* change license header
