+++ pom.xml
@@ -6,12 +6,11 @@
         <version>1.0.0</version>
     </parent>
 
-    <groupId>com.bonitasoft.connectors</groupId>
     <artifactId>bonita-connector-document-converter</artifactId>
-    <version>1.1.1-SNAPSHOT</version>
+    <version>2.0.0-SNAPSHOT</version>
 
     <properties>
-        <bonita.engine.version>7.1.3</bonita.engine.version>
+        <bonita.engine.version>7.10.5</bonita.engine.version>
         <xdocreport.version>2.0.2</xdocreport.version>
     </properties>
