--- src/main/java/com/bonitasoft/connectors/document/converter/core/DocumentConverter.java
@@ -1,18 +1,0 @@
-/*
- * Copyright (C) 2015 Bonitasoft S.A.
- * Bonitasoft is a trademark of Bonitasoft SA.
- * This software file is BONITASOFT CONFIDENTIAL. Not For Distribution.
- * For commercial licensing information, contact:
- * Bonitasoft, 32 rue Gustave Eiffel – 38000 Grenoble
- * or Bonitasoft US, 51 Federal Street, Suite 305, San Francisco, CA 94107
- */
-package com.bonitasoft.connectors.document.converter.core;
-
-import java.io.IOException;
-
-import fr.opensagres.xdocreport.core.XDocReportException;
-
-public interface DocumentConverter {
-
-    byte[] convert() throws IOException, XDocReportException;
-}
