+++ src/main/resources/document-converter.impl
@@ -5,7 +5,7 @@
 	<definitionId>document-converter</definitionId>
 	<definitionVersion>1.0.2</definitionVersion>
     <description>Implementation based xdocreport</description>
-	<implementationClassname>com.bonitasoft.connectors.document.converter.DocumentConverterConnector</implementationClassname>
+	<implementationClassname>org.bonitasoft.connectors.document.converter.DocumentConverterConnector</implementationClassname>
 	<implementationId>document-converter-impl</implementationId>
 	<implementationVersion>$implementation.version$</implementationVersion>
