+++ src/test/java/org/bonitasoft/connectors/document/converter/DocToPDFConverterTest.java
@@ -6,13 +6,12 @@
  * Bonitasoft, 32 rue Gustave Eiffel – 38000 Grenoble
  * or Bonitasoft US, 51 Federal Street, Suite 305, San Francisco, CA 94107
  */
-package com.bonitasoft.connectors.document.converter;
+package org.bonitasoft.connectors.document.converter;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.bonitasoft.connectors.document.converter.core.DocToPDFConverter;
 import org.junit.Test;
-
-import com.bonitasoft.connectors.document.converter.core.DocToPDFConverter;
 
 public class DocToPDFConverterTest {
