+++ src/test/java/org/bonitasoft/connectors/document/converter/InputParametersValidatorTest.java
@@ -6,7 +6,7 @@
  * Bonitasoft, 32 rue Gustave Eiffel – 38000 Grenoble
  * or Bonitasoft US, 51 Federal Street, Suite 305, San Francisco, CA 94107
  */
-package com.bonitasoft.connectors.document.converter;
+package org.bonitasoft.connectors.document.converter;
 
 import static org.assertj.core.api.Assertions.entry;
