+++ src/test/java/org/bonitasoft/connectors/document/converter/core/DocToHTMLConverterTest.java
@@ -6,7 +6,7 @@
  * Bonitasoft, 32 rue Gustave Eiffel – 38000 Grenoble
  * or Bonitasoft US, 51 Federal Street, Suite 305, San Francisco, CA 94107
  */
-package com.bonitasoft.connectors.document.converter.core;
+package org.bonitasoft.connectors.document.converter.core;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -14,7 +14,6 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
-import com.bonitasoft.connectors.document.converter.DocToPDFConverterTest;
 
 public class DocToHTMLConverterTest {
 
@@ -30,7 +29,7 @@ public class DocToHTMLConverterTest {
 
     @Test
     public void should_convert_docx_to_xhtml() throws Exception {
-        final DocToHTMLConverter converter = new DocToHTMLConverter(DocToPDFConverterTest.class.getResourceAsStream("/patern-signets.docx"), "utf-8");
+        final DocToHTMLConverter converter = new DocToHTMLConverter(DocToHTMLConverterTest.class.getResourceAsStream("/patern-signets.docx"), "utf-8");
 
         final byte[] content = converter.convert();
 
@@ -39,7 +38,7 @@ public class DocToHTMLConverterTest {
 
     @Test
     public void should_convert_odt_to_xhtml() throws Exception {
-        final DocToHTMLConverter converter = new DocToHTMLConverter(DocToPDFConverterTest.class.getResourceAsStream("/odtTest.odt"), "utf-8");
+        final DocToHTMLConverter converter = new DocToHTMLConverter(DocToHTMLConverterTest.class.getResourceAsStream("/odtTest.odt"), "utf-8");
 
         final byte[] content = converter.convert();
